Rootfs and SWI Build Cleanup
The rootfs and SWI packages are now stored separately from the base packages.
The base and platform packages are designed as inputs for the system integrator and are separated for use by
those building their own systems.

The rootfs, SWI (and soon to be installer) parts of the builds are productizations of those inputs.
When ONL is integrated as part of a larger project these rootfs and SWI packages are not used directly.