Add a post-merge githook to invalidate the local module manifest.
The Builder module manifest is generated once from the local tree if its missing.
When new code is merged from a remote the manifest must be regenerated in order to
reference and build the new libraries.

At setup time a post-merge hook is now added to the local repository to clear the
module manifest after merging.