New SYSTEM_COMPATIBILITY_VERSION (SCV) key
This key is designed to indicate the overall system compatibility between different installations.
For example, the system configuration in ONL2.0 is different from ONL 1.0 (partition scheme, loader/version management, etc) and attempting to boot a 1.0 switch image on a 2.0 system will not work properly.

This key exists to facilitate operational compatibility and upgrade steps. When a switch image is booted against a loader with an SCV different than it expects then a full re-install will need to be performed in order to upgrade the system before that image can run. If the SCV is different then an in-place loader-upgrade cannot be performed either.

The upgrade process will be updated to include an SCV check and perform a re-install.
The loader upgrade sequence will be predicated on SCV equality.

If the SCV is not equal after all upgrade operations then switch has the option of rebooting instead of continuing the init process.
This will be configured through the sysconfig interface.