[SWL-3475] Fix Firmware upgrade after initial ONIE upgrade.
The ONIE firware upgrade staging area is not initialized properly if we don't first boot into ONIE after
the initial upgrade. This causes our workflow of ONIE then CPLD upgrade to fail the first time.

The onie-fwpkg script does not report an error about the staging area which is why our code
is not detecting the failure to install the firmware package.