Use the new common ONIE syseeprom and machine.conf infrastructure.
The ONL platform base class supports extracting, parsing, and caching the contents of the ONIE machine.conf file
and the TLVs in the ONIE system eeprom.

The mellanox platforms that rely on this information now request it from the common baseclass and use
the common ONIE JSON parsing infrastructure to populate the onie_info structure in the SYSI implementation.