3.2 LTS Kernel Build
This kernel tracks the 3.2 LTS kernel with minimal patches to support the Broadcom IPROC.
This replaces the previous 3.2.71 patched kernel put in place temporarily to support ARM.

IPROC systems will be moved to this kernel (which will continue to track 3.2 LTS) until
IPROC support is integrated with the 3.16 LTS package.

This kernel is available for IPROC systems only.