Fix Loader Upgrade Status
When the Loader upgrade step was moved prior to platform setup the upgrade status was lost
because the platform-current symlink had not been created yet.

This simple fix places the upgrade status file in the absolute directory instead of
the symlinked directory.

The loader status and the firmware status should probably be separated anyways.
The loader status doesn't belong in the platform-specific data.