platforms: netberg: Add Aurora 420 switch
It is very similar to Aurora 420, even onlp shared object part from
Aurora 620 considered as example when porting support for 420 (see
the diff between 620 and 420 to find differences if necessary).

Main difference in onlp is that we have four LEDs (1 SYS STAT, 2 PSU
and 1 for FAN), 7 thermal (4 for two PSU, 1 for MAC, 1 for front and
1 for rear) and 6 fans (4 for board, 2 for PSU).

New 420, as well as 620 and 720 requires upstream ONIE to install
correctly. Install and boot tested with ONIE "master-201805301609".

Signed-off-by: Sergey Popovich <sergey.popovich@ordnance.co>