The SFPI layer should not implement transmitter disable for QSFP modules as part of the TX_DISABLE control.
The transmitter disable bytes are standardized by the MSA and managed by the upper layer using the dev_writeb() routines when
appropriate based on module type.

Writing the transmitter enable bytes are not appropriate for CR cables and will generate write errors so attempting to
emulate this behavior without knowing the type of module is also troublesome.