Fix bug in add_package() when the realpath of the package contains an underscore.
When a package is added to the REPO any duplicates are supposed to be removed automatically.
The (package-name, version, arch) was determined by splitting the (fullpath) name by underscore.

If an underscore was present in any of the parent directories then the split did not yield
the correct information. As a result any duplicate packages were not removed and were instead
left in the REPO.

This fix splits the package basename instead of the package path.

You may now clone your ONL worktree as JEFF_IS_STUPID and it should work now.