[Accton] Fix memory leak if PSU1 is not present
The memory leak occurs on several accton platforms. It occurs on switches
that do not have PSU1 present. The memory leak is caused due to an allocated,
but empty string being returned from the onlp_file_read_str() function. The
code calling the onlp_file_read_str() function would free the returned
allocated string only if the string was not NULL and not empty.

Signed-off-by: Peter Dobransky <peter@voltanet.io>